# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
    put:
      tags:
      - user-rest-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserExtendedDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
  /api/v1/users/me:
    get:
      tags:
      - user-rest-controller
      operationId: getCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
    put:
      tags:
      - user-rest-controller
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserExtendedDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
...
  /api/v1/images/{id}/media:
    get:
      tags:
      - image-rest-controller
      operationId: getImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - image-rest-controller
      operationId: replaceImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/users/{id}/images/:
    post:
      tags:
      - user-rest-controller
      operationId: createUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
  /api/v1/users/:
    get:
      tags:
      - user-rest-controller
      operationId: getUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserBasicDTO"
    post:
      tags:
      - user-rest-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserExtendedDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserExtendedDTO"
...
  /api/v1/users/{id}/orders:
    get:
      tags:
      - user-rest-controller
      operationId: getUserOrders
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelOrderBasicDTO"
  /api/v1/users/{id}/cart:
    get:
      tags:
      - user-rest-controller
      operationId: getUserCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderExtendedDTO"
  /api/v1/users/me/orders:
    get:
      tags:
      - user-rest-controller
      operationId: getCurrentUserOrders
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelOrderBasicDTO"
...
  /api/v1/statistics/users/{userid}:
    get:
      tags:
      - statistic-rest-controller
      operationId: getMeanTicket
      parameters:
      - name: userid
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: true
        schema:
          type: string
      - name: number
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: number
                  format: double
...
  /api/v1/users/{id}/images/{avatarId}:
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: avatarId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
    OrderBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        completed:
          type: boolean
        totalPrice:
          type: number
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        deliveryAddress:
          type: string
        deliveryNote:
          type: string
        deliveryDate:
          type: string
          format: date
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        avatar:
          $ref: "#/components/schemas/ImageDTO"
    UserExtendedDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        address:
          type: string
        creationDate:
          type: string
          format: date-time
        encodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        cart:
          $ref: "#/components/schemas/OrderBasicDTO"
        avatar:
          $ref: "#/components/schemas/ImageDTO"
...
    OrderExtendedDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        completed:
          type: boolean
        totalPrice:
          type: number
        shippingCost:
          type: number
        subtotal:
          type: number
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
          format: date
        deliveryNote:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelOrderBasicDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OrderBasicDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PagedModelUserBasicDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserBasicDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...